fix compilation warning.
This commit was SVN r11601.

+++ ompi/mca/pml/ob1/pml_ob1_sendreq.c
@@ -867,11 +867,11 @@ int mca_pml_ob1_send_request_schedule_exclusive(
 
     do {
         /* allocate remaining bytes to BTLs */
-        int32_t bytes_remaining = sendreq->req_rdma_offset -
+        size_t bytes_remaining = sendreq->req_rdma_offset -
             sendreq->req_send_offset;
         size_t prev_bytes_remaining = 0, num_fail = 0;
 
-        while(bytes_remaining > 0 &&
+        while((int32_t)bytes_remaining > 0 &&
                 (sendreq->req_pipeline_depth < mca_pml_ob1.send_pipeline_depth
                  ||
                  sendreq->req_rdma_offset < sendreq->req_send.req_bytes_packed))
